requirements.txt
google-cloud-vision
openai